requirements.txt
scikit-learn==0.24.1
numpy==1.18.2
panda==0.3.1
nltk==3.5
